+++ src/main/java/io/github/astrapi69/json/JsonToXmlExtensions.java
@@ -53,13 +53,13 @@ public final class JsonToXmlExtensions
 	{
 		if (jsonString.startsWith("["))
 		{
-			JSONArray objects = new JSONArray(jsonString);
-			return XML_HEADER + XML.toString(objects);
+			JSONArray jsonArray = new JSONArray(jsonString);
+			return XML.toString(jsonArray);
 		}
 		else
 		{
-			final JSONObject json = new JSONObject(jsonString);
-			return XML_HEADER + XML.toString(json);
+			final JSONObject jsonObject = new JSONObject(jsonString);
+			return XML.toString(jsonObject);
 		}
 	}
 
@@ -109,6 +109,60 @@ public final class JsonToXmlExtensions
 				? XML_HEADER + System.lineSeparator() + XML.toString(json, indent)
 				: XML.toString(json, indent);
 		}
+	}
+
+	/**
+	 * Transform the given json as {@link String} object to a xml as {@link String} object
+	 *
+	 * @param jsonString
+	 *            the json as {@link String} object
+	 * @param indent
+	 *            the indent is the number of spaces to add to each level of indentation
+	 * @param withHeader
+	 *            the flag that indicates if a xml header should be added
+	 * @param withRootTag
+	 *            the flag that indicates if a root tag will be added
+	 * @param rootTagName
+	 *            the name of the root tag
+	 * @return the transformed xml as {@link String} object
+	 * @throws JSONException
+	 *             if there is a syntax error in the source string or a duplicated key.
+	 */
+	public static String toXml(final String jsonString, int indent, boolean withHeader,
+		boolean withRootTag, String rootTagName) throws JSONException
+	{
+		if (jsonString.startsWith("["))
+		{
+			JSONArray jsonArray = new JSONArray(jsonString);
+			return getXmlString(withHeader, withRootTag, rootTagName,
+				XML.toString(jsonArray, indent), indent);
+		}
+		else
+		{
+			final JSONObject jsonObject = new JSONObject(jsonString);
+			return getXmlString(withHeader, withRootTag, rootTagName,
+				XML.toString(jsonObject, indent), indent);
+		}
+	}
+
+	private static String getXmlString(boolean withHeader, boolean withRootTag, String rootTagName,
+		String xmlString, int indent)
+	{
+		StringBuilder sb = new StringBuilder();
+		if (withHeader)
+		{
+			sb.append(XML_HEADER).append(System.lineSeparator());
+		}
+		if (withRootTag && rootTagName != null && !rootTagName.isEmpty())
+		{
+			sb.append("<").append(rootTagName).append(">").append(System.lineSeparator());
+		}
+		sb.append(xmlString);
+		if (withRootTag && rootTagName != null && !rootTagName.isEmpty())
+		{
+			sb.append("</").append(rootTagName).append(">");
+		}
+		return sb.toString();
 	}
 
 }