+++ src/test/java/io/github/astrapi69/json/JsonToXmlExtensionsTest.java
@@ -78,7 +78,7 @@ public class JsonToXmlExtensionsTest
 
 		jsonString = "{\"person\":{\"name\":\"Anna\",\"nickname\":\"beast\",\"gender\":\"FEMALE\",\"about\":\"Ha ha ha...\",\"married\":true},\"id\":\"23\"}";
 		actual = JsonToXmlExtensions.toXml(jsonString);
-		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><person><gender>FEMALE</gender><name>Anna</name><nickname>beast</nickname><about>Ha ha ha...</about><married>true</married></person><id>23</id>";
+		expected = "<person><gender>FEMALE</gender><name>Anna</name><nickname>beast</nickname><about>Ha ha ha...</about><married>true</married></person><id>23</id>";
 		assertEquals(actual, expected);
 
 		jsonString = ReadFileExtensions.fromFile(jsonFile);
@@ -87,17 +87,17 @@ public class JsonToXmlExtensionsTest
 
 		jsonString = ReadFileExtensions.fromFile(jsonListFile);
 		actual = JsonToXmlExtensions.toXml(jsonString);
-		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><array><person><gender>FEMALE</gender><name>Anna</name><nickname>beast</nickname><about>Ha ha ha...</about><married>true</married></person><id>23</id></array><array><person><gender>MALE</gender><name>Andreas</name><nickname>cute</nickname><about>fine person</about><married>false</married></person><id>24</id></array><array><person><gender>FEMALE</gender><name>Tatjana</name><nickname>beautiful</nickname><about>Im hot</about><married>false</married></person><id>25</id></array>";
+		expected = "<array><person><gender>FEMALE</gender><name>Anna</name><nickname>beast</nickname><about>Ha ha ha...</about><married>true</married></person><id>23</id></array><array><person><gender>MALE</gender><name>Andreas</name><nickname>cute</nickname><about>fine person</about><married>false</married></person><id>24</id></array><array><person><gender>FEMALE</gender><name>Tatjana</name><nickname>beautiful</nickname><about>Im hot</about><married>false</married></person><id>25</id></array>";
 		assertEquals(actual, expected);
 
 		jsonString = ReadFileExtensions.fromFile(jsonMapFile);
 		actual = JsonToXmlExtensions.toXml(jsonString);
-		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><1>0</1><2>0</2><3>0</3><4>0</4><5>0</5>";
+		expected = "<1>0</1><2>0</2><3>0</3><4>0</4><5>0</5>";
 		assertEquals(actual, expected);
 
 		jsonString = ReadFileExtensions.fromFile(jsonCollectionFile);
 		actual = JsonToXmlExtensions.toXml(jsonString);
-		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><array><person><gender>FEMALE</gender><name>Anna</name><nickname>beast</nickname><about>Ha ha ha...</about><married>true</married></person><id>23</id></array><array><person><gender>MALE</gender><name>Andreas</name><nickname>cute</nickname><about>fine person</about><married>false</married></person><id>24</id></array><array><person><gender>FEMALE</gender><name>Tatjana</name><nickname>beautiful</nickname><about>Im hot</about><married>false</married></person><id>25</id></array>";
+		expected = "<array><person><gender>FEMALE</gender><name>Anna</name><nickname>beast</nickname><about>Ha ha ha...</about><married>true</married></person><id>23</id></array><array><person><gender>MALE</gender><name>Andreas</name><nickname>cute</nickname><about>fine person</about><married>false</married></person><id>24</id></array><array><person><gender>FEMALE</gender><name>Tatjana</name><nickname>beautiful</nickname><about>Im hot</about><married>false</married></person><id>25</id></array>";
 		assertEquals(actual, expected);
 	}
 
@@ -165,7 +165,6 @@ public class JsonToXmlExtensionsTest
 		assertEquals(actual, expected);
 	}
 
-
 	/**
 	 * Test method for {@link JsonToXmlExtensions#toXml(String, int)}
 	 *
@@ -231,6 +230,72 @@ public class JsonToXmlExtensionsTest
 			+ "        <name>Tatjana</name>\n" + "        <nickname>beautiful</nickname>\n"
 			+ "        <about>Im hot</about>\n" + "        <married>false</married>\n"
 			+ "    </person>\n" + "    <id>25</id>\n" + "</array>\n";
+		assertEquals(actual, expected);
+	}
+
+	/**
+	 * Test method for {@link JsonToXmlExtensions#toXml(String, int)}
+	 *
+	 * @throws JSONException
+	 *             if there is a syntax error in the source string or a duplicated key
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred
+	 */
+	@Test
+	public void testToXmlStringWithIndentWithHeaderAndRootTag() throws JSONException, IOException
+	{
+		String expected;
+		String actual;
+		String jsonString;
+
+		jsonString = "{\"person\":{\"name\":\"Anna\",\"nickname\":\"beast\",\"gender\":\"FEMALE\",\"about\":\"Ha ha ha...\",\"married\":true},\"id\":\"23\"}";
+		actual = JsonToXmlExtensions.toXml(jsonString, 4, true, true, "content");
+		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" + "<content>\n" + "<person>\n"
+			+ "    <gender>FEMALE</gender>\n" + "    <name>Anna</name>\n"
+			+ "    <nickname>beast</nickname>\n" + "    <about>Ha ha ha...</about>\n"
+			+ "    <married>true</married>\n" + "</person>\n" + "<id>23</id>\n" + "</content>";
+		assertEquals(actual, expected);
+
+		jsonString = ReadFileExtensions.fromFile(jsonFile);
+		actual = JsonToXmlExtensions.toXml(jsonString, 4, true, true, "content");
+		assertEquals(actual, expected);
+
+		jsonString = ReadFileExtensions.fromFile(jsonListFile);
+		actual = JsonToXmlExtensions.toXml(jsonString, 4, true, true, "content");
+		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" + "<content>\n" + "<array>\n"
+			+ "    <person>\n" + "        <gender>FEMALE</gender>\n" + "        <name>Anna</name>\n"
+			+ "        <nickname>beast</nickname>\n" + "        <about>Ha ha ha...</about>\n"
+			+ "        <married>true</married>\n" + "    </person>\n" + "    <id>23</id>\n"
+			+ "</array>\n" + "<array>\n" + "    <person>\n" + "        <gender>MALE</gender>\n"
+			+ "        <name>Andreas</name>\n" + "        <nickname>cute</nickname>\n"
+			+ "        <about>fine person</about>\n" + "        <married>false</married>\n"
+			+ "    </person>\n" + "    <id>24</id>\n" + "</array>\n" + "<array>\n"
+			+ "    <person>\n" + "        <gender>FEMALE</gender>\n"
+			+ "        <name>Tatjana</name>\n" + "        <nickname>beautiful</nickname>\n"
+			+ "        <about>Im hot</about>\n" + "        <married>false</married>\n"
+			+ "    </person>\n" + "    <id>25</id>\n" + "</array>\n" + "</content>";
+		assertEquals(actual, expected);
+
+		jsonString = ReadFileExtensions.fromFile(jsonMapFile);
+		actual = JsonToXmlExtensions.toXml(jsonString, 4, true, true, "content");
+		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" + "<content>\n" + "<1>0</1>\n"
+			+ "<2>0</2>\n" + "<3>0</3>\n" + "<4>0</4>\n" + "<5>0</5>\n" + "</content>";
+		assertEquals(actual, expected);
+
+		jsonString = ReadFileExtensions.fromFile(jsonCollectionFile);
+		actual = JsonToXmlExtensions.toXml(jsonString, 4, true, true, "content");
+		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" + "<content>\n" + "<array>\n"
+			+ "    <person>\n" + "        <gender>FEMALE</gender>\n" + "        <name>Anna</name>\n"
+			+ "        <nickname>beast</nickname>\n" + "        <about>Ha ha ha...</about>\n"
+			+ "        <married>true</married>\n" + "    </person>\n" + "    <id>23</id>\n"
+			+ "</array>\n" + "<array>\n" + "    <person>\n" + "        <gender>MALE</gender>\n"
+			+ "        <name>Andreas</name>\n" + "        <nickname>cute</nickname>\n"
+			+ "        <about>fine person</about>\n" + "        <married>false</married>\n"
+			+ "    </person>\n" + "    <id>24</id>\n" + "</array>\n" + "<array>\n"
+			+ "    <person>\n" + "        <gender>FEMALE</gender>\n"
+			+ "        <name>Tatjana</name>\n" + "        <nickname>beautiful</nickname>\n"
+			+ "        <about>Im hot</about>\n" + "        <married>false</married>\n"
+			+ "    </person>\n" + "    <id>25</id>\n" + "</array>\n" + "</content>";
 		assertEquals(actual, expected);
 	}