+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Version 2.4-SNAPSHOT
 ADDED:
 
 - new test dependency xml-jackson-extensions in version 1.1
-- new method that takes a flag that indicates if an xml header should be added
+- new method that takes a flag that indicates if a xml header should be added
+- new method that takes a flag that indicates if a xml root tag with name should be added
 
 Version 2.3
 -------------